Refactory rpc data model and test code
* Create rpc model for daemon and wallet rpc implementations
* Refactory rpc requests and responses
* Remove unused pyi
* Refactory test code
