+++ src/cpp/daemon/py_monero_daemon_rpc.cpp
@@ -52,6 +52,7 @@
  * Parts of this file are originally copyright (c) 2012-2013 The Cryptonote developers
  */
 #include "py_monero_daemon_rpc.h"
+#include "py_monero_daemon_rpc_model.h"
 #include "utils/py_monero_utils.h"
 
 static const uint64_t MAX_REQ_SIZE = 3000000;
@@ -176,16 +177,16 @@ monero::monero_version monero_daemon_rpc::get_version() {
 bool monero_daemon_rpc::is_trusted() {
   auto res = m_rpc->send_path_request("get_height");
   check_response_status(res);
-  auto get_height_response = std::make_shared<monero_get_height_response>();
-  monero_get_height_response::from_property_tree(res, get_height_response);
+  auto get_height_response = std::make_shared<monero_get_block_result>();
+  monero_get_block_result::from_property_tree(res, get_height_response);
   return !get_height_response->m_untrusted.get();
 }
 
 uint64_t monero_daemon_rpc::get_height() {
   auto res = m_rpc->send_json_request("get_block_count");
   check_response_status(res);
-  std::shared_ptr<monero_get_block_count_result> result = std::make_shared<monero_get_block_count_result>();
-  monero_get_block_count_result::from_property_tree(res, result);
+  std::shared_ptr<monero_get_block_result> result = std::make_shared<monero_get_block_result>();
+  monero_get_block_result::from_property_tree(res, result);
   if (result->m_count == boost::none) throw std::runtime_error("Could not get height");
   return result->m_count.get();
 }
@@ -198,7 +199,7 @@ std::string monero_daemon_rpc::get_block_hash(uint64_t height) {
 
 std::shared_ptr<monero_block_template> monero_daemon_rpc::get_block_template(const std::string& wallet_address, const boost::optional<int>& reserve_size) {
   MTRACE("monero_daemon_rpc::get_block_template()");
-  auto params = std::make_shared<monero_get_block_template_params>(wallet_address, reserve_size);
+  auto params = std::make_shared<monero_get_block_params>(wallet_address, reserve_size);
   auto res = m_rpc->send_json_request("get_block_template", params);
   check_response_status(res);
   std::shared_ptr<monero_block_template> tmplt = std::make_shared<monero_block_template>();
@@ -283,9 +284,9 @@ std::vector<std::shared_ptr<monero::monero_block>> monero_daemon_rpc::get_blocks
   // fetch blocks in binary
   monero_get_blocks_by_height_request request(heights);
   auto response = m_rpc->send_binary_request(request);
-  if (response->m_binary == boost::none) throw std::runtime_error("Invalid Monero Binary response");
+  if (response.m_binary == boost::none) throw std::runtime_error("Invalid Monero Binary response");
   boost::property_tree::ptree node;
-  PyMoneroUtils::binary_blocks_to_property_tree(response->m_binary.get(), node);
+  PyMoneroUtils::binary_blocks_to_property_tree(response.m_binary.get(), node);
   check_response_status(node);
   std::vector<std::shared_ptr<monero::monero_block>> blocks;
   PyMoneroBlock::from_property_tree(node, heights, blocks);
@@ -437,7 +438,7 @@ std::shared_ptr<monero_submit_tx_result> monero_daemon_rpc::submit_tx_hex(const 
 
 void monero_daemon_rpc::relay_txs_by_hash(const std::vector<std::string>& tx_hashes) {
   MTRACE("monero_daemon_rpc::relay_txs_by_hash()");
-  auto params = std::make_shared<monero_relay_tx_params>(tx_hashes);
+  auto params = std::make_shared<monero_submit_tx_params>(tx_hashes);
   auto res = m_rpc->send_json_request("relay_tx", params);
   check_response_status(res);
 }
@@ -468,7 +469,7 @@ std::vector<std::string> monero_daemon_rpc::get_tx_pool_hashes() {
 
 void monero_daemon_rpc::flush_tx_pool(const std::vector<std::string> &hashes) {
   MTRACE("monero_daemon_rpc::flush_tx_pool()");
-  auto params = std::make_shared<monero_relay_tx_params>(hashes);
+  auto params = std::make_shared<monero_submit_tx_params>(hashes);
   auto res = m_rpc->send_json_request("flush_txpool", params);
   check_response_status(res);
 }
@@ -574,7 +575,7 @@ std::vector<std::string> monero_daemon_rpc::get_alt_block_hashes() {
   auto res = m_rpc->send_path_request("get_alt_blocks_hashes");
   check_response_status(res);
   std::vector<std::string> hashes;
-  monero_get_alt_block_hashes_response::from_property_tree(res, hashes);
+  monero_get_block_result::from_property_tree(res, hashes);
   return hashes;
 }
 
@@ -767,21 +768,21 @@ std::shared_ptr<monero::monero_block_header> monero_daemon_rpc::wait_for_next_bl
   return block_listener->m_last_header;
 }
 
-std::shared_ptr<monero_bandwith_limits_params> monero_daemon_rpc::get_bandwidth_limits() {
+std::shared_ptr<monero_bandwidth_limits> monero_daemon_rpc::get_bandwidth_limits() {
   MTRACE("monero_daemon_rpc::get_bandwidth_limits()");
   auto res = m_rpc->send_path_request("get_limit");
   check_response_status(res);
-  auto limits = std::make_shared<monero_bandwith_limits_params>();
-  monero_bandwith_limits_params::from_property_tree(res, limits);
+  auto limits = std::make_shared<monero_bandwidth_limits>();
+  monero_bandwidth_limits::from_property_tree(res, limits);
   return limits;
 }
 
-std::shared_ptr<monero_bandwith_limits_params> monero_daemon_rpc::set_bandwidth_limits(int up, int down) {
+std::shared_ptr<monero_bandwidth_limits> monero_daemon_rpc::set_bandwidth_limits(int up, int down) {
   MTRACE("monero_daemon_rpc::set_bandwidth_limits()");
-  auto limits = std::make_shared<monero_bandwith_limits_params>(up, down);
+  auto limits = std::make_shared<monero_bandwidth_limits>(up, down);
   auto res = m_rpc->send_path_request("set_limit", limits);
   check_response_status(res);
-  monero_bandwith_limits_params::from_property_tree(res, limits);
+  monero_bandwidth_limits::from_property_tree(res, limits);
   return limits;
 }
