--- src/python/monero_json_request_params.pyi
@@ -1,9 +1,0 @@
-from .serializable_struct import SerializableStruct
-
-
-class MoneroJsonRequestParams(SerializableStruct):
-    """Models a Monero JSON-RPC request parameters."""
-
-    def __init__(self) -> None:
-        """Initialize a Monero JSON-RPC request params."""
-        ...
