--- src/python/monero_path_request.pyi
@@ -1,9 +1,0 @@
-from .monero_request import MoneroRequest
-
-
-class MoneroPathRequest(MoneroRequest):
-    """Models a Monero path request."""
-
-    def __init__(self) -> None:
-        """Initialize a Monero RPC path request."""
-        ...
